normalize Build-Depends versions for android-tools packages
Since there is already so much complexity that we have to deal with, we
need to find as many ways as possible to reduce the amount of info we have
to remember to make this work.  One way of doing that is standardizing all
of the Build-Depends versions so we do not need to remember why there was
an exception that let us use an older version.  There far too much
possibility for subtle, hard to find breakage caused by building against
mismatched Build-Depends.  Google only ever builds and tests with
everything building against the _exact_ same version.

Having the versions standardized also means we can script updates easily,
since all the versions will be the same across all the packages.